Refactor test_models.py into organized test_models/ package
Split the large 1,271-line test_models.py file into a structured
test_models/ package with focused test files for each model type.

New structure:
- tests/test_models/test_base.py - BaseObject and BaseKojiObject tests
- tests/test_models/test_channel.py - Channel model tests
- tests/test_models/test_external_repo.py - ExternalRepo model tests
- tests/test_models/test_group.py - Group model tests
- tests/test_models/test_host.py - Host model tests
- tests/test_models/test_permission.py - Permission model tests
- tests/test_models/test_tag.py - Tag model tests (largest section)
- tests/test_models/test_target.py - Target model tests
- tests/test_models/test_user.py - User model tests
- tests/test_models/test_registry.py - CORE_TYPES registry tests

Benefits:
- Improved maintainability and navigation
- Better organization following project conventions
- Easier to run specific model tests
- Scales well for future model additions

All 61 tests pass after refactoring.
