--- tests/test_models.py
@@ -1,1270 +1,0 @@
-"""
-koji-habitude - test_models
-
-Unit tests for koji_habitude.models module.
-
-Author: Christopher O'Brien <obriencj@gmail.com>
-License: GNU General Public License v3
-AI-Assistant: Claude 3.5 Sonnet via Cursor
-"""
-
-# Vibe-Coding State: AI Generated
-
-import unittest
-
-from koji_habitude.models import (
-    CORE_MODELS, CORE_TYPES, BaseObject, BaseKojiObject, Channel, ExternalRepo,
-    Group, Host, Permission, Tag, Target, User,
-)
-
-
-class TestBaseModels(unittest.TestCase):
-    """
-    Test the base model classes.
-    """
-
-    def test_base_object_creation(self):
-        """
-        Test BaseObject creation with basic data.
-        """
-
-        data = {
-            'type': 'test-type',
-            'name': 'test-name',
-            '__file__': 'test.yaml',
-            '__line__': 42,
-            '__trace__': [{'key': 'value'}]
-        }
-
-        obj = BaseObject.from_dict(data)
-
-        self.assertEqual(obj.typename, 'object')
-        self.assertEqual(obj.yaml_type, 'test-type')
-        self.assertEqual(obj.name, 'test-name')
-        self.assertEqual(obj.filename, 'test.yaml')
-        self.assertEqual(obj.lineno, 42)
-        self.assertEqual(obj.trace, [{'key': 'value'}])
-        self.assertEqual(obj.data, data)
-
-    def test_base_object_key_method(self):
-        """
-        Test the key() method returns correct tuple.
-        """
-
-        data = {'type': 'test-type', 'name': 'test-name'}
-        obj = BaseObject.from_dict(data)
-
-        key = obj.key()
-        self.assertEqual(key, ('object', 'test-name'))
-
-    def test_base_object_filepos_method(self):
-        """
-        Test the filepos() method returns correct tuple.
-        """
-
-        data = {
-            'type': 'test-type',
-            'name': 'test-name',
-            '__file__': 'test.yaml',
-            '__line__': 42
-        }
-        obj = BaseObject.from_dict(data)
-
-        filepos = obj.filepos()
-        self.assertEqual(filepos, ('test.yaml', 42))
-
-    def test_base_object_can_split(self):
-        """
-        Test that BaseObject cannot be split.
-        """
-
-        data = {'type': 'test-type', 'name': 'test-name'}
-        obj = BaseObject.from_dict(data)
-
-        self.assertFalse(obj.can_split())
-
-    def test_base_object_split_raises_error(self):
-        """
-        Test that BaseObject.split() raises TypeError.
-        """
-
-        data = {'type': 'test-type', 'name': 'test-name'}
-        obj = BaseObject.from_dict(data)
-
-        with self.assertRaises(TypeError) as context:
-            obj.split()
-
-        self.assertIn('Cannot split object', str(context.exception))
-
-    def test_base_object_dependency_keys(self):
-        """
-        Test that BaseObject has no dependencies.
-        """
-
-        data = {'type': 'test-type', 'name': 'test-name'}
-        obj = BaseObject.from_dict(data)
-
-        deps = obj.dependency_keys()
-        self.assertEqual(deps, ())
-
-    def test_base_object_repr(self):
-        """
-        Test the string representation of BaseObject.
-        """
-
-        data = {'type': 'test-type', 'name': 'test-name'}
-        obj = BaseObject.from_dict(data)
-
-        repr_str = repr(obj)
-        self.assertEqual(repr_str, '<BaseObject(object, test-name)>')
-
-    def test_raw_object_alias(self):
-        """
-        Test that BaseObject is an alias for BaseObject.
-        """
-
-        data = {'type': 'test-type', 'name': 'test-name'}
-        obj = BaseObject.from_dict(data)
-
-        self.assertIsInstance(obj, BaseObject)
-        self.assertEqual(obj.typename, 'object')
-
-    def test_base_koji_object_creation(self):
-        """
-        Test BaseKojiObject creation.
-        """
-
-        data = {'type': 'koji-object', 'name': 'test-name'}
-        obj = BaseKojiObject.from_dict(data)
-
-        self.assertEqual(obj.typename, 'koji-object')
-        self.assertEqual(obj.name, 'test-name')
-
-    def test_base_koji_object_can_split_default(self):
-        """
-        Test that BaseKojiObject cannot be split by default.
-        """
-
-        data = {'type': 'koji-object', 'name': 'test-name'}
-        obj = BaseKojiObject.from_dict(data)
-
-        self.assertFalse(obj.can_split())
-
-    def test_base_koji_object_split_raises(self):
-        """
-        Test that BaseKojiObject.split() creates a minimal copy.
-        """
-
-        data = {'type': 'koji-object', 'name': 'test-name'}
-        obj = BaseKojiObject.from_dict(data)
-
-        with self.assertRaises(TypeError):
-            obj.split()
-
-    def test_base_koji_object_diff_stub(self):
-        # TODO: Implement test for BaseKojiObject.diff()
-        pass
-
-
-class TestChannelModel(unittest.TestCase):
-    """
-    Test the Channel model.
-    """
-
-    def test_channel_creation_with_defaults(self):
-        """
-        Test Channel creation with default values.
-        """
-
-        data = {'type': 'channel', 'name': 'test-channel'}
-        channel = Channel.from_dict(data)
-
-        self.assertEqual(channel.typename, 'channel')
-        self.assertEqual(channel.name, 'test-channel')
-        self.assertEqual(channel.hosts, [])
-        self.assertTrue(channel.can_split())
-
-    def test_channel_creation_with_hosts(self):
-        """
-        Test Channel creation with hosts list.
-        """
-
-        data = {
-            'type': 'channel',
-            'name': 'test-channel',
-            'hosts': ['host1', 'host2', 'host3']
-        }
-        channel = Channel.from_dict(data)
-
-        self.assertEqual(channel.hosts, ['host1', 'host2', 'host3'])
-
-    def test_channel_dependency_keys(self):
-        """
-        Test Channel dependency resolution.
-        """
-
-        data = {
-            'type': 'channel',
-            'name': 'test-channel',
-            'hosts': ['host1', 'host2']
-        }
-        channel = Channel.from_dict(data)
-
-        deps = channel.dependency_keys()
-        expected = [('host', 'host1'), ('host', 'host2')]
-        self.assertEqual(deps, expected)
-
-    def test_channel_dependency_keys_empty(self):
-        """
-        Test Channel dependency resolution with no hosts.
-        """
-
-        data = {'type': 'channel', 'name': 'test-channel'}
-        channel = Channel.from_dict(data)
-
-        deps = channel.dependency_keys()
-        self.assertEqual(deps, [])
-
-
-class TestExternalRepoModel(unittest.TestCase):
-    """
-    Test the ExternalRepo model.
-    """
-
-    def test_external_repo_creation(self):
-        """
-        Test ExternalRepo creation with valid URL.
-        """
-
-        data = {
-            'type': 'external-repo',
-            'name': 'test-repo',
-            'url': 'https://example.com/repo'
-        }
-        repo = ExternalRepo.from_dict(data)
-
-        self.assertEqual(repo.typename, 'external-repo')
-        self.assertEqual(repo.name, 'test-repo')
-        self.assertEqual(repo.url, 'https://example.com/repo')
-        self.assertFalse(repo.can_split())
-
-    def test_external_repo_creation_http_url(self):
-        """
-        Test ExternalRepo creation with HTTP URL.
-        """
-
-        data = {
-            'type': 'external-repo',
-            'name': 'test-repo',
-            'url': 'http://example.com/repo'
-        }
-        repo = ExternalRepo.from_dict(data)
-
-        self.assertEqual(repo.url, 'http://example.com/repo')
-
-    def test_external_repo_invalid_url_pattern(self):
-        """
-        Test ExternalRepo creation with invalid URL pattern.
-        """
-
-        data = {
-            'type': 'external-repo',
-            'name': 'test-repo',
-            'url': 'ftp://example.com/repo'
-        }
-
-        with self.assertRaises(ValueError):
-            ExternalRepo.from_dict(data)
-
-    def test_external_repo_dependency_keys(self):
-        """
-        Test ExternalRepo has no dependencies.
-        """
-
-        data = {
-            'type': 'external-repo',
-            'name': 'test-repo',
-            'url': 'https://example.com/repo'
-        }
-        repo = ExternalRepo.from_dict(data)
-
-        deps = repo.dependency_keys()
-        self.assertEqual(deps, ())
-
-
-class TestGroupModel(unittest.TestCase):
-    """
-    Test the Group model.
-    """
-
-    def test_group_creation_with_defaults(self):
-        """
-        Test Group creation with default values.
-        """
-
-        data = {'type': 'group', 'name': 'test-group'}
-        group = Group.from_dict(data)
-
-        self.assertEqual(group.typename, 'group')
-        self.assertEqual(group.name, 'test-group')
-        self.assertEqual(group.members, [])
-        self.assertEqual(group.permissions, [])
-        self.assertTrue(group.can_split())
-
-    def test_group_creation_with_members_and_permissions(self):
-        """
-        Test Group creation with members and permissions.
-        """
-
-        data = {
-            'type': 'group',
-            'name': 'test-group',
-            'members': ['user1', 'user2'],
-            'permissions': ['admin', 'build']
-        }
-        group = Group.from_dict(data)
-
-        self.assertEqual(group.members, ['user1', 'user2'])
-        self.assertEqual(group.permissions, ['admin', 'build'])
-
-    def test_group_dependency_keys(self):
-        """
-        Test Group dependency resolution.
-        """
-
-        data = {
-            'type': 'group',
-            'name': 'test-group',
-            'members': ['user1', 'user2'],
-            'permissions': ['admin', 'build']
-        }
-        group = Group.from_dict(data)
-
-        deps = group.dependency_keys()
-        expected = [
-            ('user', 'user1'),
-            ('user', 'user2'),
-            ('permission', 'admin'),
-            ('permission', 'build')
-        ]
-        self.assertEqual(deps, expected)
-
-    def test_group_dependency_keys_empty(self):
-        """
-        Test Group dependency resolution with no members or permissions.
-        """
-
-        data = {'type': 'group', 'name': 'test-group'}
-        group = Group.from_dict(data)
-
-        deps = group.dependency_keys()
-        self.assertEqual(deps, [])
-
-
-class TestHostModel(unittest.TestCase):
-    """
-    Test the Host model.
-    """
-
-    def test_host_creation_with_defaults(self):
-        """
-        Test Host creation with default values.
-        """
-
-        data = {'type': 'host', 'name': 'test-host'}
-        host = Host.from_dict(data)
-
-        self.assertEqual(host.typename, 'host')
-        self.assertEqual(host.name, 'test-host')
-        self.assertEqual(host.arches, [])
-        self.assertEqual(host.capacity, None)
-        self.assertTrue(host.enabled)
-        self.assertEqual(host.description, None)
-        self.assertEqual(host.channels, [])
-        self.assertTrue(host.can_split())
-
-    def test_host_creation_with_all_fields(self):
-        """
-        Test Host creation with all fields specified.
-        """
-
-        data = {
-            'type': 'host',
-            'name': 'test-host',
-            'arches': ['x86_64', 'i686'],
-            'capacity': 2.5,
-            'enabled': False,
-            'description': 'Test build host',
-            'channels': ['default', 'build']
-        }
-        host = Host.from_dict(data)
-
-        self.assertEqual(host.arches, ['x86_64', 'i686'])
-        self.assertEqual(host.capacity, 2.5)
-        self.assertFalse(host.enabled)
-        self.assertEqual(host.description, 'Test build host')
-        self.assertEqual(host.channels, ['default', 'build'])
-
-    def test_host_split(self):
-        """
-        Test Host splitting functionality.
-        """
-
-        data = {
-            'type': 'host',
-            'name': 'test-host',
-            'arches': ['x86_64'],
-            'capacity': 2.0,
-            'enabled': True,
-            'description': 'Test host',
-            'channels': ['default']
-        }
-        host = Host.from_dict(data)
-
-        split_host = host.split()
-        self.assertIsInstance(split_host, Host)
-        self.assertEqual(split_host.name, 'test-host')
-        self.assertEqual(split_host.arches, ['x86_64'])
-        self.assertEqual(split_host.capacity, 2.0)
-        self.assertTrue(split_host.enabled)
-        self.assertEqual(split_host.description, 'Test host')
-        # Channels should not be included in split (dependency data)
-        self.assertEqual(split_host.channels, [])
-
-    def test_host_dependency_keys(self):
-        """
-        Test Host dependency resolution.
-        """
-
-        data = {
-            'type': 'host',
-            'name': 'test-host',
-            'channels': ['channel1', 'channel2']
-        }
-        host = Host.from_dict(data)
-
-        deps = host.dependency_keys()
-        expected = [('channel', 'channel1'), ('channel', 'channel2')]
-        self.assertEqual(deps, expected)
-
-    def test_host_dependency_keys_empty(self):
-        """
-        Test Host dependency resolution with no channels.
-        """
-
-        data = {'type': 'host', 'name': 'test-host'}
-        host = Host.from_dict(data)
-
-        deps = host.dependency_keys()
-        self.assertEqual(deps, [])
-
-
-class TestPermissionModel(unittest.TestCase):
-    """
-    Test the Permission model.
-    """
-
-    def test_permission_creation(self):
-        """
-        Test Permission creation (name-only object).
-        """
-
-        data = {'type': 'permission', 'name': 'admin'}
-        permission = Permission.from_dict(data)
-
-        self.assertEqual(permission.typename, 'permission')
-        self.assertEqual(permission.name, 'admin')
-        self.assertFalse(permission.can_split())
-
-    def test_permission_dependency_keys(self):
-        """
-        Test Permission has no dependencies.
-        """
-
-        data = {'type': 'permission', 'name': 'admin'}
-        permission = Permission.from_dict(data)
-
-        deps = permission.dependency_keys()
-        self.assertEqual(deps, ())
-
-
-class TestTagModel(unittest.TestCase):
-    """
-    Test the Tag model.
-    """
-
-    def test_tag_creation_with_defaults(self):
-        """
-        Test Tag creation with default values.
-        """
-
-        data = {'type': 'tag', 'name': 'test-tag'}
-        tag = Tag.from_dict(data)
-
-        self.assertEqual(tag.typename, 'tag')
-        self.assertEqual(tag.name, 'test-tag')
-        self.assertEqual(tag.arches, [])
-        self.assertFalse(tag.maven_support)
-        self.assertFalse(tag.maven_include_all)
-        self.assertEqual(tag.extras, {})
-        self.assertEqual(tag.groups, {})
-        self.assertEqual(tag.inheritance, [])
-        self.assertEqual(tag.external_repos, [])
-        self.assertTrue(tag.can_split())
-
-    def test_tag_creation_with_all_fields(self):
-        """
-        Test Tag creation with all fields specified.
-        """
-
-        data = {
-            'type': 'tag',
-            'name': 'test-tag',
-            'arches': ['x86_64', 'i686'],
-            'maven-support': True,
-            'maven-include-all': True,
-            'extras': {'key1': 'value1', 'key2': 'value2'},
-            'groups': {'group1': ['pkg1', 'pkg2']},
-            'inheritance': [
-                {'name': 'parent1', 'type': 'tag', 'priority': 10},
-                {'name': 'parent2', 'type': 'tag', 'priority': 20},
-                {'name': 'repo1', 'type': 'external-repo', 'priority': 30},
-                {'name': 'repo2', 'type': 'external-repo', 'priority': 40}
-            ]
-        }
-        tag = Tag.from_dict(data)
-
-        self.assertEqual(tag.arches, ['x86_64', 'i686'])
-        self.assertTrue(tag.maven_support)
-        self.assertTrue(tag.maven_include_all)
-        self.assertEqual(tag.extras, {'key1': 'value1', 'key2': 'value2'})
-
-        # Check groups structure (now TagGroup objects instead of simple lists)
-        self.assertEqual(len(tag.groups), 1)
-        self.assertIn('group1', tag.groups)
-        group1 = tag.groups['group1']
-        self.assertEqual(group1.name, 'group1')
-        self.assertEqual(len(group1.packages), 2)
-        self.assertEqual(group1.packages[0].name, 'pkg1')
-        self.assertEqual(group1.packages[0].type, 'package')
-        self.assertEqual(group1.packages[0].block, False)
-        self.assertEqual(group1.packages[1].name, 'pkg2')
-        self.assertEqual(group1.packages[1].type, 'package')
-        self.assertEqual(group1.packages[1].block, False)
-
-        # Check inheritance links (unified list)
-        self.assertEqual(len(tag.inheritance), 4)
-
-        # Check parent tags (filtered by type)
-        parent_tags = tag.parent_tags
-        self.assertEqual(len(parent_tags), 2)
-        self.assertEqual(parent_tags[0].name, 'parent1')
-        self.assertEqual(parent_tags[0].type, 'tag')
-        self.assertEqual(parent_tags[0].priority, 10)
-        self.assertEqual(parent_tags[1].name, 'parent2')
-        self.assertEqual(parent_tags[1].type, 'tag')
-        self.assertEqual(parent_tags[1].priority, 20)
-
-        # Check external repo links (filtered by type)
-        external_repos = tag.external_repos
-        self.assertEqual(len(external_repos), 2)
-        self.assertEqual(external_repos[0].name, 'repo1')
-        self.assertEqual(external_repos[0].type, 'external-repo')
-        self.assertEqual(external_repos[0].priority, 30)
-        self.assertEqual(external_repos[1].name, 'repo2')
-        self.assertEqual(external_repos[1].type, 'external-repo')
-        self.assertEqual(external_repos[1].priority, 40)
-
-    def test_tag_split(self):
-        """
-        Test Tag splitting functionality.
-        """
-
-        data = {
-            'type': 'tag',
-            'name': 'test-tag',
-            'arches': ['x86_64'],
-            'maven-support': True,
-            'inheritance': [
-                {'name': 'parent', 'type': 'tag', 'priority': 10},
-                {'name': 'repo', 'type': 'external-repo', 'priority': 30}
-            ]
-        }
-        tag = Tag.from_dict(data)
-
-        split_tag = tag.split()
-        self.assertIsInstance(split_tag, Tag)
-        self.assertEqual(split_tag.name, 'test-tag')
-        self.assertEqual(split_tag.arches, ['x86_64'])
-        # Inheritance should not be included in split (dependency data)
-        self.assertEqual(split_tag.inheritance, [])
-        # Properties should return empty lists when inheritance is empty
-        self.assertEqual(split_tag.parent_tags, [])
-        self.assertEqual(split_tag.external_repos, [])
-
-    def test_tag_dependency_keys(self):
-        """
-        Test Tag dependency resolution.
-        """
-
-        data = {
-            'type': 'tag',
-            'name': 'test-tag',
-            'inheritance': [
-                {'name': 'parent1', 'type': 'tag', 'priority': 10},
-                {'name': 'parent2', 'type': 'tag', 'priority': 20},
-                {'name': 'repo1', 'type': 'external-repo', 'priority': 30},
-                {'name': 'repo2', 'type': 'external-repo', 'priority': 40}
-            ]
-        }
-        tag = Tag.from_dict(data)
-
-        deps = tag.dependency_keys()
-        expected = [
-            ('tag', 'parent1'),
-            ('tag', 'parent2'),
-            ('external-repo', 'repo1'),
-            ('external-repo', 'repo2')
-        ]
-        self.assertEqual(deps, expected)
-
-    def test_tag_dependency_keys_empty(self):
-        """
-        Test Tag dependency resolution with no inheritance or external repos.
-        """
-
-        data = {'type': 'tag', 'name': 'test-tag'}
-        tag = Tag.from_dict(data)
-
-        deps = tag.dependency_keys()
-        self.assertEqual(deps, [])
-
-    def test_tag_inheritance_property_filtering(self):
-        """
-        Test that parent_tags and external_repos properties correctly filter inheritance by type.
-        """
-
-        data = {
-            'type': 'tag',
-            'name': 'test-tag',
-            'inheritance': [
-                {'name': 'parent1', 'type': 'tag', 'priority': 10},
-                {'name': 'repo1', 'type': 'external-repo', 'priority': 30},
-                {'name': 'parent2', 'type': 'tag', 'priority': 20},
-                {'name': 'repo2', 'type': 'external-repo', 'priority': 40}
-            ]
-        }
-        tag = Tag.from_dict(data)
-
-        # Check that all inheritance items are present
-        self.assertEqual(len(tag.inheritance), 4)
-
-        # Check parent_tags property filters correctly
-        parent_tags = tag.parent_tags
-        self.assertEqual(len(parent_tags), 2)
-        self.assertEqual(parent_tags[0].name, 'parent1')
-        self.assertEqual(parent_tags[0].type, 'tag')
-        self.assertEqual(parent_tags[1].name, 'parent2')
-        self.assertEqual(parent_tags[1].type, 'tag')
-
-        # Check external_repos property filters correctly
-        external_repos = tag.external_repos
-        self.assertEqual(len(external_repos), 2)
-        self.assertEqual(external_repos[0].name, 'repo1')
-        self.assertEqual(external_repos[0].type, 'external-repo')
-        self.assertEqual(external_repos[1].name, 'repo2')
-        self.assertEqual(external_repos[1].type, 'external-repo')
-
-    def test_tag_inheritance_simple_string_list(self):
-        """
-        Test Tag creation with inheritance as a simple list of strings.
-        Strings should be converted to tag inheritance with incrementing priorities.
-        """
-
-        data = {
-            'type': 'tag',
-            'name': 'test-tag',
-            'inheritance': ['parent1', 'parent2', 'parent3']
-        }
-        tag = Tag.from_dict(data)
-
-        # Check that inheritance was converted to full objects
-        self.assertEqual(len(tag.inheritance), 3)
-
-        # Check first parent (priority 0)
-        self.assertEqual(tag.inheritance[0].name, 'parent1')
-        self.assertEqual(tag.inheritance[0].type, 'tag')
-        self.assertEqual(tag.inheritance[0].priority, 0)
-
-        # Check second parent (priority 10)
-        self.assertEqual(tag.inheritance[1].name, 'parent2')
-        self.assertEqual(tag.inheritance[1].type, 'tag')
-        self.assertEqual(tag.inheritance[1].priority, 10)
-
-        # Check third parent (priority 20)
-        self.assertEqual(tag.inheritance[2].name, 'parent3')
-        self.assertEqual(tag.inheritance[2].type, 'tag')
-        self.assertEqual(tag.inheritance[2].priority, 20)
-
-        # Check parent_tags property returns all items (all are type 'tag')
-        parent_tags = tag.parent_tags
-        self.assertEqual(len(parent_tags), 3)
-        self.assertEqual([p.name for p in parent_tags], ['parent1', 'parent2', 'parent3'])
-
-        # Check external_repos property returns empty (no external repos)
-        external_repos = tag.external_repos
-        self.assertEqual(len(external_repos), 0)
-
-    def test_tag_inheritance_mixed_string_and_dict_list(self):
-        """
-        Test Tag creation with inheritance mixing strings and dict objects.
-        Strings should be converted to tag inheritance, dicts should be used as-is.
-        """
-
-        data = {
-            'type': 'tag',
-            'name': 'test-tag',
-            'inheritance': [
-                'parent1',  # String - should become tag with priority 0
-                {'name': 'parent2', 'type': 'tag', 'priority': 50},  # Dict with explicit priority
-                'parent3',  # String - should become tag with priority 10
-                {'name': 'repo1', 'type': 'external-repo', 'priority': 25}  # Dict with different type
-            ]
-        }
-        tag = Tag.from_dict(data)
-
-        # Check that inheritance was processed correctly
-        self.assertEqual(len(tag.inheritance), 4)
-
-        # Check string conversion (parent1)
-        self.assertEqual(tag.inheritance[0].name, 'parent1')
-        self.assertEqual(tag.inheritance[0].type, 'tag')
-        self.assertEqual(tag.inheritance[0].priority, 0)
-
-        # Check dict with explicit priority (parent2)
-        self.assertEqual(tag.inheritance[1].name, 'parent2')
-        self.assertEqual(tag.inheritance[1].type, 'tag')
-        self.assertEqual(tag.inheritance[1].priority, 50)
-
-        # Check string conversion (parent3) - priority should increment from max existing
-        self.assertEqual(tag.inheritance[2].name, 'parent3')
-        self.assertEqual(tag.inheritance[2].type, 'tag')
-        self.assertEqual(tag.inheritance[2].priority, 60)  # 50 + 10
-
-        # Check dict with different type (repo1)
-        self.assertEqual(tag.inheritance[3].name, 'repo1')
-        self.assertEqual(tag.inheritance[3].type, 'external-repo')
-        self.assertEqual(tag.inheritance[3].priority, 25)
-
-        # Check property filtering
-        parent_tags = tag.parent_tags
-        self.assertEqual(len(parent_tags), 3)
-        self.assertEqual([p.name for p in parent_tags], ['parent1', 'parent2', 'parent3'])
-
-        external_repos = tag.external_repos
-        self.assertEqual(len(external_repos), 1)
-        self.assertEqual(external_repos[0].name, 'repo1')
-
-    def test_tag_inheritance_empty_list(self):
-        """
-        Test Tag creation with empty inheritance list.
-        """
-
-        data = {
-            'type': 'tag',
-            'name': 'test-tag',
-            'inheritance': []
-        }
-        tag = Tag.from_dict(data)
-
-        # Check that inheritance is empty
-        self.assertEqual(len(tag.inheritance), 0)
-        self.assertEqual(len(tag.parent_tags), 0)
-        self.assertEqual(len(tag.external_repos), 0)
-
-    def test_tag_groups_dict_format(self):
-        """
-        Test Tag creation with groups in dict format: {'group_name': ['pkg1', 'pkg2']}
-        """
-
-        data = {
-            'type': 'tag',
-            'name': 'test-tag',
-            'groups': {'build': ['package1', 'package2'], 'srpm': ['source1']}
-        }
-        tag = Tag.from_dict(data)
-
-        # Check groups structure
-        self.assertEqual(len(tag.groups), 2)
-
-        # Check build group
-        self.assertIn('build', tag.groups)
-        build_group = tag.groups['build']
-        self.assertEqual(build_group.name, 'build')
-        self.assertEqual(len(build_group.packages), 2)
-        self.assertEqual(build_group.packages[0].name, 'package1')
-        self.assertEqual(build_group.packages[0].type, 'package')
-        self.assertEqual(build_group.packages[0].block, False)
-        self.assertEqual(build_group.packages[1].name, 'package2')
-        self.assertEqual(build_group.packages[1].type, 'package')
-        self.assertEqual(build_group.packages[1].block, False)
-
-        # Check srpm group
-        self.assertIn('srpm', tag.groups)
-        srpm_group = tag.groups['srpm']
-        self.assertEqual(srpm_group.name, 'srpm')
-        self.assertEqual(len(srpm_group.packages), 1)
-        self.assertEqual(srpm_group.packages[0].name, 'source1')
-        self.assertEqual(srpm_group.packages[0].type, 'package')
-        self.assertEqual(srpm_group.packages[0].block, False)
-
-    def test_tag_groups_list_format(self):
-        """
-        Test Tag creation with groups in list format: [{'name': 'group', 'packages': ['pkg']}]
-        """
-
-        data = {
-            'type': 'tag',
-            'name': 'test-tag',
-            'groups': [
-                {'name': 'build', 'packages': ['package1', 'package2']},
-                {'name': 'srpm', 'packages': ['source1']}
-            ]
-        }
-        tag = Tag.from_dict(data)
-
-        # Check groups structure
-        self.assertEqual(len(tag.groups), 2)
-
-        # Check build group
-        self.assertIn('build', tag.groups)
-        build_group = tag.groups['build']
-        self.assertEqual(build_group.name, 'build')
-        self.assertEqual(len(build_group.packages), 2)
-        self.assertEqual(build_group.packages[0].name, 'package1')
-        self.assertEqual(build_group.packages[1].name, 'package2')
-
-        # Check srpm group
-        self.assertIn('srpm', tag.groups)
-        srpm_group = tag.groups['srpm']
-        self.assertEqual(srpm_group.name, 'srpm')
-        self.assertEqual(len(srpm_group.packages), 1)
-        self.assertEqual(srpm_group.packages[0].name, 'source1')
-
-    def test_tag_groups_dict_with_package_objects(self):
-        """
-        Test Tag creation with groups containing package objects with type and block.
-        """
-
-        data = {
-            'type': 'tag',
-            'name': 'test-tag',
-            'groups': {
-                'build': {
-                    'packages': [
-                        'package1',  # Simple string
-                        {'name': 'package2', 'type': 'package', 'block': False},
-                        {'name': '@group1', 'type': 'group', 'block': True}
-                    ]
-                }
-            }
-        }
-        tag = Tag.from_dict(data)
-
-        # Check groups structure
-        self.assertEqual(len(tag.groups), 1)
-
-        # Check build group
-        build_group = tag.groups['build']
-        self.assertEqual(build_group.name, 'build')
-        self.assertEqual(len(build_group.packages), 3)
-
-        # First package (simple string)
-        self.assertEqual(build_group.packages[0].name, 'package1')
-        self.assertEqual(build_group.packages[0].type, 'package')
-        self.assertEqual(build_group.packages[0].block, False)
-
-        # Second package (explicit dict)
-        self.assertEqual(build_group.packages[1].name, 'package2')
-        self.assertEqual(build_group.packages[1].type, 'package')
-        self.assertEqual(build_group.packages[1].block, False)
-
-        # Third package (group with @ prefix - name should keep the @ when specified explicitly)
-        self.assertEqual(build_group.packages[2].name, '@group1')
-        self.assertEqual(build_group.packages[2].type, 'group')
-        self.assertEqual(build_group.packages[2].block, True)
-
-    def test_tag_groups_string_package_with_at_prefix(self):
-        """
-        Test Tag creation with string packages that have @ prefix (indicating group type).
-        """
-
-        data = {
-            'type': 'tag',
-            'name': 'test-tag',
-            'groups': {
-                'build': ['package1', '@group1', 'package2', '@group2']
-            }
-        }
-        tag = Tag.from_dict(data)
-
-        # Check groups structure
-        build_group = tag.groups['build']
-        self.assertEqual(len(build_group.packages), 4)
-
-        # First package (no @ prefix)
-        self.assertEqual(build_group.packages[0].name, 'package1')
-        self.assertEqual(build_group.packages[0].type, 'package')
-        self.assertEqual(build_group.packages[0].block, False)
-
-        # Second package (@ prefix)
-        self.assertEqual(build_group.packages[1].name, '@group1')
-        self.assertEqual(build_group.packages[1].type, 'group')
-        self.assertEqual(build_group.packages[1].block, False)
-
-        # Third package (no @ prefix)
-        self.assertEqual(build_group.packages[2].name, 'package2')
-        self.assertEqual(build_group.packages[2].type, 'package')
-        self.assertEqual(build_group.packages[2].block, False)
-
-        # Fourth package (@ prefix)
-        self.assertEqual(build_group.packages[3].name, '@group2')
-        self.assertEqual(build_group.packages[3].type, 'group')
-        self.assertEqual(build_group.packages[3].block, False)
-
-    def test_tag_groups_duplicate_group_names_error(self):
-        """
-        Test that duplicate group names in list format raise an error.
-        """
-
-        data = {
-            'type': 'tag',
-            'name': 'test-tag',
-            'groups': [
-                {'name': 'build', 'packages': ['package1']},
-                {'name': 'build', 'packages': ['package2']}  # Duplicate name
-            ]
-        }
-
-        with self.assertRaises(TypeError) as context:
-            Tag.from_dict(data)
-
-        self.assertIn('Duplicate group build', str(context.exception))
-
-    def test_tag_groups_invalid_string_in_dict_error(self):
-        """
-        Test that string values in dict format raise an error.
-        """
-
-        data = {
-            'type': 'tag',
-            'name': 'test-tag',
-            'groups': {
-                'build': 'package1'  # String instead of list/dict
-            }
-        }
-
-        with self.assertRaises(ValueError) as context:
-            Tag.from_dict(data)
-
-        self.assertIn('Group build must be a dictionary or list', str(context.exception))
-
-    def test_tag_groups_invalid_data_type_error(self):
-        """
-        Test that invalid data types for groups raise an error.
-        """
-
-        data = {
-            'type': 'tag',
-            'name': 'test-tag',
-            'groups': 'invalid'  # String instead of dict/list
-        }
-
-        with self.assertRaises(ValueError) as context:
-            Tag.from_dict(data)
-
-        self.assertIn('Groups must be a dictionary or list', str(context.exception))
-
-
-class TestTargetModel(unittest.TestCase):
-    """
-    Test the Target model.
-    """
-
-    def test_target_creation_with_build_tag(self):
-        """
-        Test Target creation with build-tag specified.
-        """
-
-        data = {
-            'type': 'target',
-            'name': 'test-target',
-            'build-tag': 'build-tag'
-        }
-        target = Target.from_dict(data)
-
-        self.assertEqual(target.typename, 'target')
-        self.assertEqual(target.name, 'test-target')
-        self.assertEqual(target.build_tag, 'build-tag')
-        self.assertEqual(target.dest_tag, None)
-        self.assertFalse(target.can_split())
-
-    def test_target_creation_with_both_tags(self):
-        """
-        Test Target creation with both build-tag and dest-tag specified.
-        """
-
-        data = {
-            'type': 'target',
-            'name': 'test-target',
-            'build-tag': 'build-tag',
-            'dest-tag': 'dest-tag'
-        }
-        target = Target.from_dict(data)
-
-        self.assertEqual(target.build_tag, 'build-tag')
-        self.assertEqual(target.dest_tag, 'dest-tag')
-
-    def test_target_model_post_init_sets_dest_tag(self):
-        """
-        Test that model_post_init sets dest_tag to name when not specified.
-        """
-
-        data = {
-            'type': 'target',
-            'name': 'test-target',
-            'build-tag': 'build-tag'
-            # dest-tag not specified
-        }
-        target = Target.from_dict(data)
-
-        self.assertEqual(target.dest_tag, None)
-
-    def test_target_dependency_keys(self):
-        """
-        Test Target dependency resolution.
-        """
-
-        data = {
-            'type': 'target',
-            'name': 'test-target',
-            'build-tag': 'build-tag',
-            'dest-tag': 'dest-tag'
-        }
-        target = Target.from_dict(data)
-
-        deps = target.dependency_keys()
-        expected = [('tag', 'build-tag'), ('tag', 'dest-tag')]
-        self.assertEqual(deps, expected)
-
-
-class TestUserModel(unittest.TestCase):
-    """
-    Test the User model.
-    """
-
-    def test_user_creation_with_defaults(self):
-        """
-        Test User creation with default values.
-        """
-
-        data = {'type': 'user', 'name': 'test-user'}
-        user = User.from_dict(data)
-
-        self.assertEqual(user.typename, 'user')
-        self.assertEqual(user.name, 'test-user')
-        self.assertEqual(user.groups, [])
-        self.assertEqual(user.permissions, [])
-        self.assertTrue(user.enabled)
-        self.assertTrue(user.can_split())
-
-    def test_user_creation_with_all_fields(self):
-        """
-        Test User creation with all fields specified.
-        """
-
-        data = {
-            'type': 'user',
-            'name': 'test-user',
-            'groups': ['group1', 'group2'],
-            'permissions': ['admin', 'build'],
-            'enabled': False
-        }
-        user = User.from_dict(data)
-
-        self.assertEqual(user.groups, ['group1', 'group2'])
-        self.assertEqual(user.permissions, ['admin', 'build'])
-        self.assertFalse(user.enabled)
-
-    def test_user_split(self):
-        """
-        Test User splitting functionality.
-        """
-
-        data = {
-            'type': 'user',
-            'name': 'test-user',
-            'groups': ['group1', 'group2'],
-            'permissions': ['admin', 'build'],
-            'enabled': False
-        }
-        user = User.from_dict(data)
-        split_user = user.split()
-        self.assertIsInstance(split_user, User)
-        self.assertEqual(split_user.name, 'test-user')
-        self.assertEqual(split_user.groups, [])
-        self.assertEqual(split_user.permissions, [])
-        self.assertFalse(split_user.enabled)
-
-    def test_user_dependency_keys(self):
-        """
-        Test User dependency resolution.
-        """
-
-        data = {
-            'type': 'user',
-            'name': 'test-user',
-            'groups': ['group1', 'group2'],
-            'permissions': ['admin', 'build']
-        }
-        user = User.from_dict(data)
-
-        deps = user.dependency_keys()
-        expected = [
-            ('group', 'group1'),
-            ('group', 'group2'),
-            ('permission', 'admin'),
-            ('permission', 'build')
-        ]
-        self.assertEqual(deps, expected)
-
-    def test_user_dependency_keys_empty(self):
-        """
-        Test User dependency resolution with no groups or permissions.
-        """
-
-        data = {'type': 'user', 'name': 'test-user'}
-        user = User.from_dict(data)
-
-        deps = user.dependency_keys()
-        self.assertEqual(deps, [])
-
-
-class TestCoreModelsRegistry(unittest.TestCase):
-    """
-    Test the CORE_TYPES registry and model consistency.
-    """
-
-    def test_core_models_contains_all_expected_models(self):
-        """
-        Test that CORE_TYPES contains all expected model classes.
-        """
-
-        expected_models = [
-            Channel,
-            ExternalRepo,
-            Group,
-            Host,
-            Permission,
-            Tag,
-            Target,
-            User,
-        ]
-
-        self.assertEqual(len(CORE_TYPES), len(expected_models))
-
-        for model_class in expected_models:
-            self.assertIn(model_class, CORE_TYPES)
-
-    def test_core_models_have_correct_typenames(self):
-        """
-        Test that all CORE_TYPES have correct typename attributes.
-        """
-
-        expected_typenames = {
-            Channel: 'channel',
-            ExternalRepo: 'external-repo',
-            Group: 'group',
-            Host: 'host',
-            Permission: 'permission',
-            Tag: 'tag',
-            Target: 'target',
-            User: 'user',
-        }
-
-        for model_class in CORE_TYPES:
-            expected_typename = expected_typenames[model_class]
-            self.assertEqual(model_class.typename, expected_typename)
-
-    def test_core_models_can_split_support(self):
-        """
-        Test that CORE_TYPES have correct can_split support.
-        """
-
-        expected_split_support = {
-            Channel: True,
-            ExternalRepo: False,
-            Group: True,
-            Host: True,
-            Permission: False,
-            Tag: True,
-            Target: False,
-            User: True,
-        }
-
-        for model_class in CORE_TYPES:
-            expected_support = expected_split_support[model_class]
-            self.assertEqual(model_class._can_split, expected_support)
-
-    def test_core_models_instantiation_with_minimal_data(self):
-        """
-        Test that all CORE_TYPES can be instantiated with minimal data.
-        """
-
-        minimal_data_templates = {
-            Channel: {'type': 'channel', 'name': 'test'},
-            ExternalRepo: {'type': 'external-repo', 'name': 'test', 'url': 'https://example.com'},
-            Group: {'type': 'group', 'name': 'test'},
-            Host: {'type': 'host', 'name': 'test'},
-            Permission: {'type': 'permission', 'name': 'test'},
-            Tag: {'type': 'tag', 'name': 'test'},
-            Target: {'type': 'target', 'name': 'test', 'build-tag': 'build-tag'},
-            User: {'type': 'user', 'name': 'test'},
-        }
-
-        for model_class in CORE_TYPES:
-            data = minimal_data_templates[model_class]
-            obj = model_class.from_dict(data)
-            self.assertIsInstance(obj, BaseKojiObject)
-            self.assertEqual(obj.name, 'test')
-
-    def test_core_models_dependency_keys_return_tuples(self):
-        """
-        Test that all CORE_TYPES dependency_keys() methods return proper tuples.
-        """
-
-        minimal_data_templates = {
-            Channel: {'type': 'channel', 'name': 'test'},
-            ExternalRepo: {'type': 'external-repo', 'name': 'test', 'url': 'https://example.com'},
-            Group: {'type': 'group', 'name': 'test'},
-            Host: {'type': 'host', 'name': 'test'},
-            Permission: {'type': 'permission', 'name': 'test'},
-            Tag: {'type': 'tag', 'name': 'test'},
-            Target: {'type': 'target', 'name': 'test', 'build-tag': 'build-tag'},
-            User: {'type': 'user', 'name': 'test'},
-        }
-
-        for model_class in CORE_TYPES:
-            data = minimal_data_templates[model_class]
-            obj = model_class.from_dict(data)
-            deps = obj.dependency_keys()
-
-            self.assertIsInstance(deps, (list, tuple))
-            for dep in deps:
-                self.assertIsInstance(dep, tuple)
-                self.assertEqual(len(dep), 2)
-                self.assertIsInstance(dep[0], str)  # type
-                self.assertIsInstance(dep[1], str)  # name
-
-
-# The end.
