+++ tests/test_models/__init__.py
@@ -1,0 +1,11 @@
+"""
+koji-habitude - test_models
+
+Test package for koji_habitude.models module.
+
+Author: Christopher O'Brien <obriencj@gmail.com>
+License: GNU General Public License v3
+AI-Assistant: Claude 3.5 Sonnet via Cursor
+"""
+
+# The end.
